requirements.txt
asgiref
Django
django-cors-headers
django-environ
djangorestframework
pillow
sqlparse
tzdata
